+++ CHANGELOG.md
@@ -34,6 +34,8 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
    See [setStickerSetThumbnail](https://core.telegram.org/bots/api#setstickersetthumbnail) for more information.
  * Added method `\TgBotLib\Bot > setStickerSetTitle()` to set the title of a sticker set. 
    See [setStickerSetTitle](https://core.telegram.org/bots/api#setstickersettitle) for more information.
+ * Added method `\TgBotLib\Bot > deleteStickerSet()` to delete a sticker set. 
+   See [deleteStickerSet](https://core.telegram.org/bots/api#deletestickerset) for more information.
 
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`